# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Intelligent AI Assistant API
  description: |
    Next-generation **Agentic AI Life Assistant** API designed to act as the user's conscious digital twin, 
    continuously learning, guiding, and optimizing real-world actions in real time.
    
    ## Features
    - **Voice Processing**: Speech-to-text and text-to-speech with Whisper and Coqui/ElevenLabs
    - **Calendar Integration**: Google Calendar synchronization with intelligent scheduling
    - **WhatsApp Messaging**: Business Cloud API integration for confirmations and updates
    - **Federated Learning**: Privacy-preserving machine learning with FedAvg and Differential Privacy
    - **Digital Avatar**: 2D/3D animated user representation during interactions
    - **Privacy Compliance**: GDPR, DIFC, and UAE Data Protection Law compliance
    
    ## Authentication
    All endpoints (except health checks and webhook verification) require JWT authentication.
    Include the Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - **Authentication endpoints**: 5 requests per minute
    - **Voice processing**: 20 requests per minute
    - **General API**: 100 requests per minute
    
    ## Error Handling
    The API uses standard HTTP status codes and returns structured error responses:
    ```json
    {
      "error": "Error description",
...
      "request_id": "uuid",
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /voice/stt:
    post:
      tags: [Voice Processing]
      summary: Speech to text
      description: |
        Convert speech audio to text using Whisper STT.
        
        **Audio Requirements:**
        - Format: WAV, MP3, M4A, FLAC
        - Max size: 50MB
        - Base64 encoded
        
        **Performance:**
        - Average latency: < 2 seconds
        - Accuracy: > 95% for clear speech
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [audio_data]
              properties:
                audio_data:
                  type: string
                  format: base64
                  description: Base64 encoded audio data
                  example: UklGRnoGAABXQVZFZm10IBAAAAABAAEAQB8AAEAfAAABAAgAZGF0YQoGAACBhYqFbF1fdJivrJBhNjVgodDbq2EcBj+a2/LDciUFLIHO8tiJNwgZaLvt559NEAxQp+PwtmMcBjiR1/LMeSwFJHfH8N2QQAoUXrTp66hVFApGn+DyvmwhBSuBzvLZiTYIG2m98OScTgwOUarm7blmGgU7k9n1unEiBC13yO/eizEIHWq+8+OWT...
                language:
                  type: string
                  description: Language hint (ISO 639-1 code)
                  example: en
                session_id:
                  type: string
                  description: Optional session ID for context
                  example: session_123
      responses:
        '200':
          description: Transcription successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
                    example: "Schedule a meeting with John tomorrow at 3 PM"
                  confidence:
                    type: number
                    format: float
                    minimum: 0
                    maximum: 1
                    example: 0.95
                  language:
                    type: string
                    example: en
                  processing_time:
                    type: number
                    format: float
                    example: 1.23
                  session_id:
                    type: string
                    example: session_123
        '400':
          description: Invalid audio data or format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /voice/tts:
    post:
      tags: [Voice Processing]
      summary: Text to speech
      description: |
        Convert text to speech using Coqui TTS or ElevenLabs.
        
        **Text Requirements:**
        - Max length: 5000 characters
        - Supports multiple languages
        
        **Audio Output:**
        - Format: WAV (16kHz, 16-bit)
        - Quality: High-fidelity synthesis
        - Personalization: User voice characteristics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  maxLength: 5000
                  example: "Your meeting with John has been scheduled for tomorrow at 3 PM"
                language:
                  type: string
                  default: en
                  example: en
                voice_id:
                  type: string
                  description: Specific voice ID (optional)
                  example: voice_user_123
                prefer_quality:
                  type: boolean
                  default: false
                  description: Prefer quality over speed
                speed:
                  type: number
                  format: float
                  minimum: 0.5
                  maximum: 2.0
                  default: 1.0
                  description: Speech speed multiplier
      responses:
        '200':
          description: TTS synthesis successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  audio_url:
                    type: string
                    format: uri
                    example: "/api/v1/voice/audio/audio_123.wav"
                  audio_data:
                    type: string
                    format: base64
                    description: Base64 audio data (for small files)
                    nullable: true
                  metadata:
                    type: object
                    properties:
                      duration:
                        type: number
                        format: float
                        example: 3.45
                      sample_rate:
                        type: integer
                        example: 16000
                      voice_engine:
                        type: string
                        example: coqui
                      processing_time:
                        type: number
                        format: float
                        example: 0.89
        '400':
          description: Invalid text or parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Calendar Integration Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.
        
        **Token Format:**
        ```
        Authorization: Bearer <jwt_token>
        ```
        
        **Token Lifecycle:**
        - Access tokens expire after 1 hour
        - Refresh tokens expire after 30 days
        - Automatic token rotation supported
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Invalid email or password"
        correlation_id:
          type: string
          format: uuid
          description: Request correlation ID for debugging
          example: "550e8400-e29b-41d4-a716-446655440000"
        request_id:
          type: string
          format: uuid
          description: Unique request identifier
          example: "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
        timestamp:
          type: number
          format: float
          description: Unix timestamp of error
          example: 1640995200.0
        details:
          type: object
          description: Additional error details
          additionalProperties: true
...